+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/bisq_easy/open_trades/message_box/MyProtocolLogMessageBox.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.desktop.main.content.bisq_easy.open_trades;
+package bisq.desktop.main.content.bisq_easy.open_trades.message_box;
 
 import bisq.chat.ChatChannel;
 import bisq.chat.ChatMessage;
