+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/chat/message_container/list/ChatMessageListCellFactory.java
@@ -20,8 +20,8 @@ package bisq.desktop.main.content.chat.message_container.list;
 import bisq.chat.ChatChannel;
 import bisq.chat.ChatMessage;
 import bisq.chat.bisqeasy.open_trades.BisqEasyOpenTradeMessage;
-import bisq.desktop.main.content.bisq_easy.open_trades.MyProtocolLogMessageBox;
-import bisq.desktop.main.content.bisq_easy.open_trades.PeerProtocolLogMessageBox;
+import bisq.desktop.main.content.bisq_easy.open_trades.message_box.MyProtocolLogMessageBox;
+import bisq.desktop.main.content.bisq_easy.open_trades.message_box.PeerProtocolLogMessageBox;
 import bisq.desktop.main.content.chat.message_container.list.message_box.*;
 import javafx.geometry.Insets;
 import javafx.geometry.Pos;
