+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/chat/message_container/list/message_box/PeerLeftMessageBox.java
@@ -20,7 +20,7 @@ package bisq.desktop.main.content.chat.message_container.list.message_box;
 import bisq.chat.ChatChannel;
 import bisq.chat.ChatMessage;
 import bisq.desktop.common.utils.ImageUtil;
-import bisq.desktop.main.content.bisq_easy.open_trades.PeerProtocolLogMessageBox;
+import bisq.desktop.main.content.bisq_easy.open_trades.message_box.PeerProtocolLogMessageBox;
 import bisq.desktop.main.content.chat.message_container.list.ChatMessageListItem;
 import bisq.desktop.main.content.chat.message_container.list.ChatMessagesListController;
 import bisq.i18n.Res;
